// frontend/src/app/api/models/dia-itinerario.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TextoEnriquecido } from '../models/texto-enriquecido';
export interface DiaItinerario {
  actividades: TextoEnriquecido;
  alojamiento_orientativo?: (string | null);
  consejos?: (string | null);
  desnivel_negativo_m?: (number | null);
  desnivel_positivo_m?: (number | null);
  distancia_km?: (number | null);
  numero_dia: number;
  titulo: string;
}