// frontend/src/app/api/models/itinerario-detalle.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContenidoRelacionado } from '../models/contenido-relacionado';
import { DestinoTarjeta } from '../models/destino-tarjeta';
import { DiaItinerario } from '../models/dia-itinerario';
import { Dificultad } from '../models/dificultad';
import { FuentePublica } from '../models/fuente-publica';
import { ImagenPublica } from '../models/imagen-publica';
import { MetadatosEditoriales } from '../models/metadatos-editoriales';
import { RefTaxonomia } from '../models/ref-taxonomia';
import { Slug } from '../models/slug';
import { TerminoRef } from '../models/termino-ref';
import { TextoEnriquecido } from '../models/texto-enriquecido';
export interface ItinerarioDetalle {
  desnivel_acumulado_m?: (number | null);
  destino: DestinoTarjeta;
  dias: Array<DiaItinerario>;
  dificultad: Dificultad;
  distancia_total_km?: (number | null);
  duracion_dias: number;
  fuentes: Array<FuentePublica>;
  galeria: Array<ImagenPublica>;
  metadatos: MetadatosEditoriales;
  portada: ImagenPublica;
  relacionados: Array<ContenidoRelacionado>;
  resumen: string;
  riesgos_seguridad: TextoEnriquecido;
  slug: Slug;
  terminos: Array<TerminoRef>;
  tipos: Array<RefTaxonomia>;
  titulo: string;
}